Move AddOnPriority/PaymentStatus into Order namespace
