+++ src/Forms/CheckoutForm.php
@@ -21,8 +21,8 @@ use SilverStripe\Omnipay\Model\Message\PurchaseError;
 use SilverStripe\Omnipay\Model\Payment;
 use SilverStripe\Omnipay\Service\ServiceFactory;
 use SilverStripe\Security\Security;
-use SwipeStripe\Constants\PaymentStatus;
 use SwipeStripe\Order\Order;
+use SwipeStripe\Order\PaymentStatus;
 use SwipeStripe\Pages\OrderConfirmationPage;
 use SwipeStripe\Price\SupportedCurrencies\SupportedCurrenciesInterface;
