+++ src/Order/OrderAddOn.php
@@ -8,7 +8,6 @@ use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\FieldType\DBInt;
 use SilverStripe\ORM\FieldType\DBVarchar;
 use SilverStripe\Versioned\Versioned;
-use SwipeStripe\Constants\AddOnPriority;
 use SwipeStripe\Price\DBPrice;
 use SwipeStripe\Price\PriceField;
